requirements.txt
django_heroku
bs4
pytube
gunicorn